Rename base components
