+++ src/qua_dashboards/video_mode/tab_controllers/__init__.py
@@ -1,9 +1,9 @@
-from .itab_controller import ITabController
+from .base_tab_controller import BaseTabController
 from .live_view_tab_controller import LiveViewTabController
 from .annotation_tab_controller import AnnotationTabController
 
 __all__ = [
-    "ITabController",
+    "BaseTabController",
     "LiveViewTabController",
     "AnnotationTabController",
 ]
