+++ src/qua_dashboards/video_mode/utils/__init__.py
@@ -1,3 +1,3 @@
-from .annotation_tools import *
+from .annotation_utils import *
 
-__all__ = annotation_tools.__all__
+__all__ = annotation_utils.__all__
