+++ src/qua_dashboards/video_mode/utils/annotation_utils.py
